fix(facts): cursor_item_id widened to BigInteger, revision f3a9c41d2b07 amended
cursor_item_id copies media_items.id, which is BigInteger. The brief justified
Integer by analogy to TmdbRateState, but that analogy only licenses the pinned
id=1 primary key beside it -- a row's own identity, answerable by inspection.
A column that stores a foreign value inherits that value's width; the question
is media_items', and it is already answered.

Live ids also track upsert ATTEMPTS rather than row count: the pipeline's
on_conflict_do_update (render/pipeline.py:309-312) burns a sequence value on
every re-processed item, because Postgres evaluates nextval before the
conflict check. That is very likely why media_items.id is BigInteger already.

The revision is unshipped -- no deployed table to ALTER -- so it is amended
rather than followed up. Both edits land together or alembic check reports
model-vs-migration drift. The up/down/up proof was re-run in-container: its
column assertion now pins (id, integer), (cursor_item_id, bigint),
(updated_at, timestamptz), and the round-trip row inserts 2**31+7 -- a value
the old integer column would have refused.

+++ alembic/versions/f3a9c41d2b07_facts_backfill_state.py
@@ -23,7 +23,7 @@ def upgrade() -> None:
     op.create_table(
         'facts_backfill_state',
         sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('cursor_item_id', sa.Integer(), nullable=True),
+        sa.Column('cursor_item_id', sa.BigInteger(), nullable=True),
         sa.Column(
             'updated_at', sa.DateTime(timezone=True),
             server_default=sa.text('now()'), nullable=False,
